app/requirements.txt
flask
openai
azure-identity
azure-keyvault-secrets
pandas
numpy
transformers